+++ src/components/Sidebar/SidebarPanel/SidebarPanel.vue
@@ -33,6 +33,7 @@ const areFiltersChosen = computed(() => {
     return filters.value.includeSunsetWeapons
         || filters.value.craftedWeapons
         || filters.value.adeptWeapons
+        || !!filters.value.perkFilter
         || filters.value.collectionsFilters.length > 0
         || filters.value.damageFilters.length > 0
         || filters.value.rarityFilters.length > 0
@@ -47,15 +48,20 @@ const filteredWeapons = computed(() => {
     }
 
     const filtered = weapons.value
-        .filter(w => !filters.value.includeSunsetWeapons || !w.isSunset)
+        .filter(w => filters.value.includeSunsetWeapons || !w.isSunset)
         .filter(w => !filters.value.craftedWeapons || w.isCraftable)
         .filter(w => !filters.value.adeptWeapons || w.isAdept)
+        .filter(w => filters.value.perkFilter ? filters.value.perkFilter(w) : true)
         .filter(w => checkFilterCategoryOnWeapon(filters.value.collectionsFilters, w))
         .filter(w => checkFilterCategoryOnWeapon(filters.value.damageFilters, w))
         .filter(w => checkFilterCategoryOnWeapon(filters.value.rarityFilters, w))
         .filter(w => checkFilterCategoryOnWeapon(filters.value.weaponFilters, w))
         .filter(w => w.name.toLocaleLowerCase().includes(props.searchString.toLocaleLowerCase()));
-    console.log("filtered weapons", filtered);
+    const test1 = weapons.value.filter(w => !w.isSunset);
+    const test2 = weapons.value.filter(w => filters.value.perkFilter ? filters.value.perkFilter(w) : true);
+    const test3 = test2.filter(w => !w.isSunset);
+    const test4 = test1.filter(w => filters.value.perkFilter ? filters.value.perkFilter(w) : true);
+    console.log("filters are", filters.value, test1, test2, test3, test4);
     return filtered;
 });